Use ColumnCell as React.memo
